+++ code/go/0chain.net/smartcontract/dbs/event/event.go
@@ -59,19 +59,6 @@ func (edb *EventDb) GetEvents(block int64) ([]Event, error) {
 	return events, result.Error
 }
 
-func (edb *EventDb) drop() error {
-	err := edb.Store.Get().Migrator().DropTable(&Event{})
-	if err != nil {
-		return err
-	}
-
-	err = edb.Store.Get().Migrator().DropTable(&Blobber{})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 func (edb *EventDb) exists(event Event) (bool, error) {
 	var count int64
 	result := edb.Store.Get().
