+++ code/go/0chain.net/smartcontract/dbs/postgresql/postresql.go
@@ -76,49 +76,3 @@ func (store *PostgresStore) Close() {
 func (store *PostgresStore) Get() *gorm.DB {
 	return store.db
 }
-
-/*
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
- */
